delete post javascript

+++ cms/admin/includes/posts_view-all-posts.php
@@ -101,7 +101,7 @@ if (isset($_POST['checkBoxArray'])) {
 				echo "<td>{$post_date}</td>";
 				echo "<td><a href='../post.php?p_id={$post_id}'>View Post</a></td>";
 				echo "<td><a href='./posts.php?source=edit_post&p_id={$post_id}'>edit</a></td>";
-				echo "<td><a href='./posts.php?delete={$post_id}'>delete</a></td>";
+				echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete?');\" href='./posts.php?delete={$post_id}'>delete</a></td>";
 				echo "</tr>";
 			}
 			?>
